backend/requirements.txt
# -----------------------------
# Web framework
# -----------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# -----------------------------
# Streaming (SSE)
# -----------------------------
sse-starlette>=1.8.2

# -----------------------------
# Data validation
# -----------------------------
pydantic>=2.6.0

# -----------------------------
# Environment variables
# -----------------------------
python-dotenv>=1.0.1

# -----------------------------
# LLM client (Perplexity via OpenAI-compatible SDK)
# -----------------------------
openai>=1.12.0

# -----------------------------
# Async HTTP (used internally)
# -----------------------------
httpx>=0.27.0
anyio>=4.2.0